fix: Move mock server
